moves all logic to CreateTurboVotePostInRogue
